+++ src/components/core.ts
@@ -5,6 +5,7 @@ import * as _ from './utils';
 import { LogLevels } from './utils';
 import { EditorConfig, OutputData, SanitizerConfig } from '../../types';
 import { EditorModules } from '../types-internal/editor-modules';
+import I18n from './i18n';
 
 /**
  * @typedef {Core} Core - editor core class
@@ -202,6 +203,13 @@ export default class Core {
       if (!this.config.data.blocks || this.config.data.blocks.length === 0) {
         this.config.data.blocks = [ initialBlockData ];
       }
+    }
+
+    /**
+     * Adjust i18n
+     */
+    if (config.i18n && config.i18n.messages) {
+      I18n.setDictionary(config.i18n.messages);
     }
   }