+++ types/api/index.d.ts
@@ -11,3 +11,4 @@ export * from './notifier';
 export * from './tooltip';
 export * from './inline-toolbar';
 export * from './block';
+export * from './i18n';